+++ lending/src/models/components/items.ts
@@ -13,27 +13,15 @@ export type Items = {
 
 /** @internal */
 export namespace Items$ {
-    export const inboundSchema: z.ZodType<Items, z.ZodTypeDef, unknown> = z
-        .object({
-            id: z.string().optional(),
-        })
-        .transform((v) => {
-            return {
-                ...(v.id === undefined ? null : { id: v.id }),
-            };
-        });
+    export const inboundSchema: z.ZodType<Items, z.ZodTypeDef, unknown> = z.object({
+        id: z.string().optional(),
+    });
 
     export type Outbound = {
         id?: string | undefined;
     };
 
-    export const outboundSchema: z.ZodType<Outbound, z.ZodTypeDef, Items> = z
-        .object({
-            id: z.string().optional(),
-        })
-        .transform((v) => {
-            return {
-                ...(v.id === undefined ? null : { id: v.id }),
-            };
-        });
+    export const outboundSchema: z.ZodType<Outbound, z.ZodTypeDef, Items> = z.object({
+        id: z.string().optional(),
+    });
 }